daedalus-core/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
groq==0.11.0
python-dotenv==1.0.1
httpx==0.27.2
websockets==13.1
supabase==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
ruff==0.7.0
